move helper functions out of middleware files
...since there is no longer custom middleware.
