+++ src/components/Form/TimeoutWarning/TimeoutWarning.jsx
@@ -3,12 +3,10 @@ import { connect } from 'react-redux'
 import { i18n } from '../../../config'
 import { api } from '../../../services'
 import { updateApplication } from '../../../actions/ApplicationActions'
-import { logout, tokenError } from '../../../actions/AuthActions'
-import { saveSection } from '../../../middleware/history'
+import { tokenError } from '../../../actions/AuthActions'
+import { saveSection } from '../../SavedIndicator/persistence-helpers'
 import AuthenticatedView from '../../../views/AuthenticatedView'
-import Branch from '../Branch'
 import Modal from '../Modal'
-import Show from '../Show'
 import Svg from '../Svg'
 
 export const roundUp = (num, precision) => {
