+++ src/views/Form/Form.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import { withRouter } from 'react-router'
 import { api } from '../../services'
-import { findPosition, push, saveSection } from '../../middleware/history'
+import { saveSection } from '../../components/SavedIndicator/persistence-helpers'
 import { clearErrors, updateApplication } from '../../actions/ApplicationActions'
 import AuthenticatedView from '../AuthenticatedView'
 import { Section, SavedIndicator, TimeoutWarning } from '../../components'
 import { env } from '../../config'
-import { parseFormUrl } from '../../components/Navigation/navigation-helpers'
+import { findPosition, parseFormUrl } from '../../components/Navigation/navigation-helpers'
 import { tokenError } from '../../actions/AuthActions'
 import { unstickAll } from '../../components/Sticky/sidebar'
 import { updateSection } from '../../actions/SectionActions'
